Pass only the correct data to each of the owner fields.
This fixes the problem of the form failing validation, but it is still broken
in the sense that once an owner is set for either of the fields, it is impossible to
remove all owners from that field. This is because if the multienum is empty,
the default value is applied to it when it is submitted.